app/requirements.txt
Django>=1.10,<1.11
Fabric==1.12.0
psycopg2>=2.6,<2.7
ipython>=5.1.0
django-grappelli>=2.8,<2.9
django-simple-rest>=1.4,<1.5
Pillow==3.3.0
RPi.GPIO==0.5.11
Spidev==3.0